新增RuntimeObjectManager用以实现运行时对象的管理。实现EngineCore的接口。目前GetRuntimeInfo未完全实现，GetRuntimeComponent未实现，后期再实现。

+++ source/src/Modules/Core/MasterCore/EngineHandle.cs
@@ -85,12 +85,12 @@ namespace Testflow.MasterCore
             return _runtimeEngine.GetRuntimeInfo<TDataType>(infoName, extraParams);
         }
 
-        public int AddRuntimeObject(string objectType, int sessionId, params object[] param)
+        public long AddRuntimeObject(string objectType, int sessionId, params object[] param)
         {
             return _runtimeEngine.AddRuntimeObject(objectType, sessionId, param);
         }
 
-        public int RemoveRuntimeObject(int objectId, params object[] param)
+        public long RemoveRuntimeObject(int objectId, params object[] param)
         {
             return _runtimeEngine.RemoveRuntimeObject(objectId, param);
         }
